+++ templates/vst-simple/source/project1_synthesizer.h
@@ -15,7 +15,7 @@ public:
   ~Project1Synthesizer() override = default;
 
   void setupParameters(ParameterBuilder &builder) override;
-  void setParameter(uint32_t address, double value) override;
+  void setParameter(uint64_t address, double value) override;
   void prepare(double sampleRate, int32_t maxFrameCount) override;
   void noteOn(int32_t noteNumber, double velocity) override;
   void noteOff(int32_t noteNumber) override;
